+++ src/information.ts
