--- src/status/index.ts
@@ -1,5 +1,0 @@
-export * from './information'
-export * from './success'
-export * from './redirect'
-export * from './client_error'
-export * from './server_error'
